backend/pyproject.toml
[project]
name = "selecta"
version = "0.1.0"
description = "Selecta: backend-only Google ADK agent for BigQuery."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "Selecta Project" }
]
dependencies = [
  "google-adk>=0.4.0",
  "google-cloud-bigquery>=3.25.0",
  "google-cloud-dataplex>=2.10.0",
  "google-genai>=1.13.0",
  "google-auth>=2.27.0",
  "google-cloud-core>=2.4.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
  "proto-plus>=1.23.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app", "selecta", "selecta.*"]

[tool.setuptools.package-data]
"selecta" = ["instructions.yaml", "datasets/*.yaml"]

[tool.uv]
